+++ imsdk/src/main/java/com/masonsoft/imsdk/core/IMActionMessageManager.java
@@ -192,7 +192,7 @@ public class IMActionMessageManager {
             @NonNull
             private final IMActionMessage mActionMessage;
 
-            public long mErrorCode;
+            public int mErrorCode;
             public String mErrorMessage;
 
             private final AtomicBoolean mBuildActionMessagePacket = new AtomicBoolean(false);
@@ -416,7 +416,7 @@ public class IMActionMessageManager {
                                 }
 
                                 if (result.getCode() != 0) {
-                                    setErrorCode(result.getCode());
+                                    setErrorCode((int) result.getCode());
                                     setErrorMessage(result.getMsg());
                                     IMLog.e(Objects.defaultObjectTag(this) +
                                             " unexpected. errorCode:%s, errorMessage:%s", result.getCode(), result.getMsg());
@@ -496,7 +496,7 @@ public class IMActionMessageManager {
                                 }
 
                                 if (result.getCode() != 0) {
-                                    setErrorCode(result.getCode());
+                                    setErrorCode((int) result.getCode());
                                     setErrorMessage(result.getMsg());
                                     IMLog.e(Objects.defaultObjectTag(this) +
                                             " unexpected. errorCode:%s, errorMessage:%s", result.getCode(), result.getMsg());
